Everything is finally working. Just the position of the resize arrows on the coupled intop could be improved, but they work.

+++ model/item.cc
@@ -132,7 +132,7 @@ namespace minsky
       }          
      
     // then, check whether a resize handle has been selected  
-	double w=0.5*width()*zoomFactor(), h=0.5*height()*zoomFactor(); 
+	float w=0.5*width()*zoomFactor(), h=0.5*height()*zoomFactor(); 
     if (abs(abs(x-this->x())-w) < portRadiusMult*zoomFactor() &&	  
             abs(abs(y-this->y())-h) < portRadiusMult*zoomFactor() &&	  
             abs(hypot((x-this->x()),(y-this->y()))-hypot(w,h)) < portRadiusMult*zoomFactor())
